kafka-cdc-lab/requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
faker==20.1.0